Edit Budget controller's List action behaviour
Instead of returning List view only with budgets side menu and blank content, now we display the details of first budget in the returned list of budgets from database. If database doesn't return any budgets, a page with blank content will be displayed with avaiable options.

We don't display previous budgets anymore. A separate function will be added later for that purpose.